CDAP-13654 remove unused dataset fields from predictor plugin
Also removing some classes that were leftover from when training
took place in a plugin instead of a service.
