--- mmds-plugins/widgets/ModelTrainer-sparksink.json
@@ -1,144 +1,0 @@
-{
-  "metadata": {
-    "spec-version": "1.4"
-  },
-  "configuration-groups": [
-    {
-      "label": "Model Configuration",
-      "properties": [
-        {
-          "widget-type": "textbox",
-          "label": "Experiment ID",
-          "name": "experimentId"
-        },
-        {
-          "widget-type": "textbox",
-          "label": "Model ID",
-          "name": "modelId"
-        },
-        {
-          "widget-type": "textbox",
-          "label": "Outcome Field",
-          "name": "outcomeField"
-        },
-        {
-          "widget-type": "select",
-          "label": "Outcome Is Categorical",
-          "name": "outcomeIsCategorical",
-          "widget-attributes": {
-            "values": [
-              "false",
-              "true"
-            ],
-            "default": "false"
-          }
-        },
-        {
-          "widget-type": "select",
-          "label": "Algorithm",
-          "name": "algorithm",
-          "widget-attributes": {
-            "values": [
-              "linear.regression",
-              "generalized.linear.regression",
-              "decision.tree.regression",
-              "random.forest.regression",
-              "gradient.boosted.tree.regression",
-              "logistic.regression",
-              "naive.bayes",
-              "decision.tree.classifier",
-              "random.forest.classifier",
-              "gradient.boosted.tree.classifier",
-              "multilayer.perceptron.classifier"
-            ]
-          }
-        },
-        {
-          "widget-type": "json-editor",
-          "label": "Training Parameters",
-          "name": "trainingParameters"
-        }
-      ]
-    },
-    {
-      "label": "Feature Configuration",
-      "properties": [
-        {
-          "widget-type": "csv",
-          "label": "Features to Include",
-          "name": "featureWhitelist",
-          "widget-attributes": {
-            "delimiter": ",",
-            "value-placeholder": "Field Name"
-          }
-        },
-        {
-          "widget-type": "csv",
-          "label": "Features to Exclude",
-          "name": "featureBlacklist",
-          "widget-attributes": {
-            "delimiter": ",",
-            "value-placeholder": "Field Name"
-          }
-        },
-        {
-          "widget-type": "csv",
-          "label": "Categorical Features",
-          "name": "categoricalFeatures",
-          "widget-attributes": {
-            "delimiter": ",",
-            "value-placeholder": "Field Name"
-          }
-        }
-      ]
-    },
-    {
-      "label": "Predictions Dataset Configuration",
-      "properties": [
-        {
-          "widget-type": "textbox",
-          "label": "Predictions Dataset",
-          "name": "predictionsDataset"
-        },
-        {
-          "widget-type": "textbox",
-          "label": "Predictions Dataset Delimiter",
-          "name": "predictionsDatasetDelimiter"
-        }
-      ]
-    },
-    {
-      "label": "Advanced Configuration",
-      "properties": [
-        {
-          "widget-type": "select",
-          "label": "Overwrite Existing",
-          "name": "overwrite",
-          "widget-attributes": {
-            "values": [
-              "false",
-              "true"
-            ],
-            "default": "false"
-          }
-        },
-        {
-          "widget-type": "textbox",
-          "label": "Test Split Percentage",
-          "name": "testSplitPercentage"
-        },
-        {
-          "widget-type": "textbox",
-          "label": "Model Dataset",
-          "name": "modelDataset"
-        },
-        {
-          "widget-type": "textbox",
-          "label": "Model MetaData Dataset",
-          "name": "modelMetaDataset"
-        }
-      ]
-    }
-  ],
-  "outputs": []
-}
